Update patch set 2
Patch Set 2: Code-Review+2

Looks good to me: we need to do some testing though.
Change A => Change B

if you make Change A submittable, it should then trigger the submission of Change B as well, as they are on the same project and branch.

Similarly, if Change C is in conflict with his target branch (not submittable) but then the conflict is resolved on the target branch, the autosubmitter should then submit Change C as well.

Patch-set: 2
Reviewer: Luca Milanesio <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2