Add end-to-end test on cross-repo below independent review
The merge of a cross-repo below an independent review is a use case we
partially support: the cross-repo can be merged but a "Code-Review -1"
is sometimes pushed on the cross-repo (this occurs when pushing a new
version of a patchset of the cross-repo). The next commit will avoid
this "Code-Review -1".

Change-Id: I920a075b81b9448010f68bd43bcc11c3f4146052