Fix submission of changes using the botUser identity
Since Change-Id: I39fe14d7e the submission isn't done anymore
using the botUser, which is exactly against the main purpose of this
plugin.

Restore the proper identity of the change submitter as botUser and add
validation tests to make sure that this fundamental feature does not
break anymore by accident and unnoticed.

Bug: Issue 10173
Change-Id: I064468385548c1696898158bd8153165684d5be2