Evaluating Submit Requirements as well as label functions.
When evaluating the submittability of a change, the plugin was just
considering the legacy submit records resulting from the evaluation of
label functions.

Since Gerrit 3.6 however, the standard and default way of defining the
submittability of a change is through the definition of Submit
Requirements, which were completely ignored.

The result was that the autosubmitter plugin was attempting to merge
changes that were not submittable at all.

Consider both submit records resulting from the evaluation of label
functions and submit requirements, so that the result match the actual
submittability status of the change.

Bug: Issue 458294128
Change-Id: I4df85941ecf518ab2875eeb1ef2a02c0bd4be1b5